+++ Examples/BankingExample/Program.cs
@@ -1,5 +1,7 @@
 ﻿using BankingExample;
 
+Console.OutputEncoding = System.Text.Encoding.UTF8;
+
 // Run all banking examples
 await BankingExamples.RunExamplesAsync();